pyproject.toml
[tool.poetry]
name = "llama-index-playground"
version = "0.1.0"
description = ""
authors = ["vikyw89 <vikyw89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
llama-index = "^0.9.20"
python-dotenv = "^1.0.0"
pypdf = "^3.17.3"
docarray = {extras = ["hnswlib"], version = "^0.40.0"}
rank-bm25 = "^0.2.2"
matplotlib = "^3.8.2"
ipython = "^8.19.0"
tiktoken = "^0.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
